+++ app/Http/Controllers/TrackController.php
@@ -6,7 +6,11 @@ use App\Models\Track;
 use App\Http\Requests\StoreTrackRequest;
 use App\Http\Requests\UpdateTrackRequest;
 use App\Models\Album;
+use App\Models\composer;
+use App\Models\EpisodeTrack;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class TrackController extends Controller
 {
@@ -17,7 +21,11 @@ class TrackController extends Controller
      */
     public function index()
     {
-        //
+        try {
+            return view('app.tracks.index-track');
+        } catch (Exception $ex) {
+            return  $this->pageError($ex);
+        }
     }
 
     /**
@@ -27,8 +35,13 @@ class TrackController extends Controller
      */
     public function create()
     {
-        $albums = Album::all();
-        return view('createtrack', compact('albums'));
+        try {
+            $albums = Album::all();
+            $composers = composer::all();
+            return view('app.tracks.createtrack', compact('albums', 'composers'));
+        } catch (\Exception $ex) {
+            return  $this->pageError($ex);
+        }
     }
 
     /**
@@ -47,16 +60,16 @@ class TrackController extends Controller
             } else {
                 $collection['active'] = false;
             }
-            $time = explode(':',$collection['length']);
+            if(!empty($collection['length'])){
+             $time = explode(':',$collection['length']);
             $collection['length'] = (int)$time[0]*60 + (int)$time[1];
-
+            }
             Track::create($collection->toArray());
-            return redirect()->back();
+            return redirect()->route('track.index');
         } catch (\Exception $ex) {
-            dd($ex);
+            return $this->pageError($ex);
         }
     }
-
     /**
      * Display the specified resource.
      *
@@ -65,7 +78,11 @@ class TrackController extends Controller
      */
     public function show(Track $track)
     {
-        return view('viewtrack', compact('track'));
+        try {
+            return view('app.tracks.viewtrack', compact('track'));
+        } catch (\Exception $ex) {
+            return $this->pageError($ex);
+        }
     }
 
     /**
@@ -78,9 +95,10 @@ class TrackController extends Controller
     {
         try {
             $albums = Album::all();
-            return view('edittrack', compact('albums', 'track'));
+            $composers = composer::all();
+            return view('app.tracks.edittrack', compact('albums', 'track','composers'));
         } catch (\Exception $ex) {
-            redirect()->route('track.index');
+            return $this->pageError($ex);
         }
     }
 
@@ -106,7 +124,7 @@ class TrackController extends Controller
             $track->update($collection->toArray());
             return redirect()->route('track.view', $track->id);
         } catch (\Exception $ex) {
-            dd($ex);
+            return $this->pageError($ex);
         }
     }
 
@@ -119,10 +137,63 @@ class TrackController extends Controller
     public function destroy(Track $track)
     {
         try {
+            if (EpisodeTrack::where('track_id', $track->id)->exists()) {
+                throw new Exception('Can not delete which has dependencies');
+            };
             $track->delete();
-            return redirect('/');
+            return redirect()->route('track.index');
         } catch (\Exception $ex) {
+            return $this->pageError($ex);
+        }
+    }
+
+    public function getRecords()
+    {
+        try {
+            $tracks = Track::all();
+            return $tracks->load(['album','composer' ])->toJson();
+        } catch (\Exception $ex) {
+            ddd($ex);
+            return $this->pageError($ex);
+        }
+    }
+
+    public function checkOrder(Request $request)
+    {
+        try {
+            $validator = Validator::make($request->all(), [
+                'order' => 'required|numeric',
+                'track' => 'exists:tracks,id',
+                'disk' => 'numeric',
+            ]);
+            if ($validator->fails()) {
+                return ['result' => '500'];
+            }
+
+            $album = Track::where('order', $request->order)->where('disk', $request->disk);
+            dd($album->get());
+            $response = '';
+            if (!$album->exists()) {
+                $response = ['result' => '1', 'data' => ''];
+            } else if ($album->first()->id == $request->album) {
+                $response =  ['result' => '2', 'data' => 'the id belongs to this album'];
+            } else {
+
+                $response =  ['result' => '0', 'data' => $album->first()->title];
+            }
+            return $response;
+        } catch (\Exception $ex) {
+            return $this->pageError($ex);
             dd($ex);
         }
+    }
+
+
+
+
+
+    private function pageError(Exception $ex)
+    {
+        return redirect()->route('error')->withError($ex->getMessage());
     }
 }
